EwramData_unk7864 cleanup/doc, EwramData_unkA098 cleanup, misc

+++ include/code/code_080211F0.h
@@ -14,7 +14,7 @@ extern void sub_080214DC(struct EwramData_unk4E4 *param_0);
 extern void sub_08021E50(s32 param_0);
 extern void sub_08021F18(void);
 extern s32 sub_080220CC(void);
-extern u16 sub_08023424(void); // Return type?
+extern s32 sub_08023424(void); // Return type looks s16, but outputs s32. May be implicitly called?
 extern void sub_080235C8(void);
 extern s32 sub_0802D5EC(s32 param_0);
 extern void sub_0802DFF0(void);
